Move queue-helpers to separate package
