+++ .github/dependabot.yml
@@ -84,3 +84,10 @@ updates:
       interval: 'weekly'
     target-branch: 'main'
     open-pull-requests-limit: 5
+
+  - package-ecosystem: 'npm'
+    directory: '/packages/queue-helpers'
+    schedule:
+      interval: 'weekly'
+    target-branch: 'main'
+    open-pull-requests-limit: 5
