+++ apps/backend/local.settings.example.json
@@ -1,15 +1,17 @@
 {
-    "IsEncrypted": false,
-    "Values": {
-      "AzureWebJobsStorage": "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;",
-      "FUNCTIONS_WORKER_RUNTIME": "node",
-      "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
-      "API_SHARED_KEY": "openssl rand -base64 32", // Should be the same in the apps/frontend
-      "COSMOSDB_ENDPOINT": "https://localhost:8081",
-      "COSMOSDB_KEY": "Your CosmosDB Key",
-      "COSMOSDB_DATABASE": "running",
-      "NODE_TLS_REJECT_UNAUTHORIZED": "0", // Only for local development
-      "NEXT_PUBLIC_STRAVA_CLIENT_ID": "Your Strava Client ID",
-      "STRAVA_CLIENT_SECRET": "Your Strava Client Secret"
-    }
+  "IsEncrypted": false,
+  "Values": {
+    "AzureWebJobsStorage": "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;",
+    "FUNCTIONS_WORKER_RUNTIME": "node",
+    "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
+    "API_SHARED_KEY": "openssl rand -base64 32", // Should be the same in the apps/frontend
+    "COSMOSDB_ENDPOINT": "https://localhost:8081",
+    "COSMOSDB_KEY": "Your CosmosDB Key",
+    "COSMOSDB_DATABASE": "running",
+    "NODE_TLS_REJECT_UNAUTHORIZED": "0", // Only for local development
+    "NEXT_PUBLIC_STRAVA_CLIENT_ID": "Your Strava Client ID",
+    "STRAVA_CLIENT_SECRET": "Your Strava Client Secret",
+    "STRAVA_DAILY_LIMIT": "1000",
+    "STRAVA_15MIN_LIMIT": "100"
   }
+}
