+++ apps/backend/package.json
@@ -15,9 +15,9 @@
   },
   "dependencies": {
     "@azure/functions": "^4.0.0",
-    "@azure/storage-queue": "12.16.0",
     "@repo/cosmosdb": "*",
     "@repo/jwt": "*",
+    "@repo/queue-helpers": "*",
     "@repo/strava": "*",
     "durable-functions": "3.0.0-alpha.5"
   },
