+++ packages/queue-helpers/index.ts
@@ -1,0 +1,4 @@
+import { addActivitiesToQueue } from './src/addActivitiesToQueue';
+import { createQueueClient } from './src/helpers';
+
+export { createQueueClient, addActivitiesToQueue };
