+++ packages/queue-helpers/readme.md
@@ -1,0 +1,9 @@
+# Add activities to a azure storage queue
+
+This package contains a helper function to add activities to a azure storage queue.
+
+## Usage
+
+```typescript
+import { addActivitiesToQueue } from '@repo/queue-helpers';
+```
