+++ packages/queue-helpers/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@repo/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": ".",
+  },
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules"]
+}
